# [SPARK-52850][PYTHON] Skip calling conversions if identity function